requirements.txt
# AI-Powered Vacation Planner Requirements
# Google Agent Development Kit (ADK) Implementation

# Core ADK
google-generativeai>=0.5.0
google-adk>=0.1.0

# Google Cloud (for Vertex AI production deployment)
google-cloud-aiplatform>=1.38.0

# External APIs
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Logging and monitoring
loguru>=0.7.0

# Document generation
python-docx>=0.8.11

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Weather API
# Install if using OpenWeather API integration

# Optional: Amadeus API
# Install if using Amadeus flight/hotel APIs
amadeus>=9.0.0
